[C]: add rest of image poll tests and change casting in bounded polls to account for signage.

+++ aeron-client/src/main/c/aeron_image.c
@@ -289,7 +289,7 @@ int aeron_image_bounded_poll(
     const uint8_t *term_buffer = image->log_buffer->mapped_raw_log.term_buffers[index].addr;
     const int32_t initial_offset = (int32_t)initial_position & image->term_length_mask;
     const int32_t capacity = (const int32_t)image->log_buffer->mapped_raw_log.term_length;
-    const int32_t high_limit_offset = (int32_t)(limit_position - initial_position + initial_offset);
+    const int64_t high_limit_offset = limit_position - initial_position + initial_offset;
     const int32_t limit_offset = capacity < high_limit_offset ? capacity : high_limit_offset;
     int32_t offset = initial_offset;
 
@@ -356,7 +356,7 @@ int aeron_image_bounded_controlled_poll(
     const uint8_t *term_buffer = image->log_buffer->mapped_raw_log.term_buffers[index].addr;
     const int32_t capacity = (const int32_t)image->log_buffer->mapped_raw_log.term_length;
     int32_t initial_offset = (int32_t)initial_position & image->term_length_mask;
-    const int32_t high_limit_offset = (int32_t)(limit_position - initial_position + initial_offset);
+    const int64_t high_limit_offset = limit_position - initial_position + initial_offset;
     const int32_t limit_offset = capacity < high_limit_offset ? capacity : high_limit_offset;
     int32_t offset = initial_offset;
 
@@ -456,7 +456,7 @@ int64_t aeron_image_controlled_peek(
     const uint8_t *term_buffer = image->log_buffer->mapped_raw_log.term_buffers[index].addr;
     const int32_t capacity = (const int32_t)image->log_buffer->mapped_raw_log.term_length;
     int32_t initial_offset = (int32_t)initial_position & image->term_length_mask;
-    const int32_t high_limit_offset = (int32_t)(limit_position - initial_position + initial_offset);
+    const int64_t high_limit_offset = limit_position - initial_position + initial_offset;
     const int32_t limit_offset = capacity < high_limit_offset ? capacity : high_limit_offset;
     int32_t offset = initial_offset;
 
@@ -538,7 +538,7 @@ int aeron_image_block_poll(
     const uint8_t *term_buffer = image->log_buffer->mapped_raw_log.term_buffers[index].addr;
     const int32_t offset = (int32_t)position & image->term_length_mask;
     const int32_t capacity = (const int32_t)image->log_buffer->mapped_raw_log.term_length;
-    const int32_t high_limit_offset = (int32_t)(offset + block_length_limit);
+    const int64_t high_limit_offset = offset + block_length_limit;
     const int32_t limit_offset = capacity < high_limit_offset ? capacity : high_limit_offset;
     int32_t scan_offset = offset;
